remove unused `resource_for_federation`
This is now only used in `test_typing`, so move it there.
